v0.7.14.c.12 Add dynamic display of orders in the basket

+++ src/main/java/dev/hotdeals/treecreate/controller/ProfileController.java
@@ -145,18 +145,25 @@ public class ProfileController
         {
             LOGGER.info("User is logging in, searching for ID: " + session.getId());
             User sessionUser = userRepo.findOneByEmail(session.getId());
-            LOGGER.info("Found user: " + sessionUser);
-            List<TreeOrder> orderList = treeOrderRepo.findAllByUserId(sessionUser.getId());
-            LOGGER.info("Found all orders for the user. Orders: " + orderList.size());
-
-            for (TreeOrder order : orderList)
+            if (sessionUser != null)
             {
-                LOGGER.info("Order " + order.getOrderId() + " is being assigned to " + user);
-                order.setUserByUserId(user);
-                treeOrderRepo.save(order);
+                LOGGER.info("Found user: " + sessionUser);
+                List<TreeOrder> orderList = treeOrderRepo.findAllByUserId(sessionUser.getId());
+                LOGGER.info("Found all orders for the user. Orders: " + orderList.size());
+
+                for (TreeOrder order : orderList)
+                {
+                    LOGGER.info("Order " + order.getOrderId() + " is being assigned to " + user);
+                    order.setUserByUserId(user);
+                    treeOrderRepo.save(order);
+                }
+                LOGGER.info("All orders have been remapped, deleting the temp user");
+                userRepo.delete(sessionUser);
+            } else
+            {
+                LOGGER.info("No session user found, moving on");
             }
-            LOGGER.info("All orders have been remapped, deleting the temp user");
-            userRepo.delete(sessionUser);
+
 
             session.setAttribute("userId", user.getId());
             LOGGER.info("New login: " + session.getId() + " as user: " + user.getId());
